pyproject.toml
[project]
name = "repomind"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "code-splitter>=0.1.5",
    "gitpython>=3.1.46",
    "langchain>=1.2.15",
    "langchain-text-splitters>=1.1.1",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
    "tree-sitter>=0.25.2",
    "tree-sitter-languages>=1.10.2",
    "tree-sitter-python>=0.25.0",
]
